fixed clang-tidy ....

+++ tasks/urin_o_edge_img_sobel/mpi/src/ops_mpi.cpp
@@ -180,8 +180,8 @@ int UrinOEdgeImgSobelMPI::GradientX(int x, int y) {
       if (nx >= 0 && nx < width_ && ny >= 0 && ny < local_height_with_halo_) {
         int pixel = local_pixels_[(static_cast<size_t>(ny) * width_) + nx];
         // const int kernel_value = kSobelX[static_cast<size_t>(ky + 1)][static_cast<size_t>(kx + 1)];
-        const auto sobel_x_idx = static_cast<size_t>(kx + 1);
-        const auto sobel_y_idx = static_cast<size_t>(ky + 1);
+        const auto sobel_x_idx = static_cast<size_t>(static_cast<long long>(kx) + 1LL);
+        const auto sobel_y_idx = static_cast<size_t>(static_cast<long long>(ky) + 1LL);
         const int kernel_value = kSobelXArray.at(sobel_y_idx).at(sobel_x_idx);
         sum += pixel * kernel_value;
       }
@@ -202,8 +202,8 @@ int UrinOEdgeImgSobelMPI::GradientY(int x, int y) {
       if (nx >= 0 && nx < width_ && ny >= 0 && ny < local_height_with_halo_) {
         int pixel = local_pixels_[(static_cast<size_t>(ny) * width_) + nx];
         // const int kernel_value = kSobelY[static_cast<size_t>(ky + 1)][static_cast<size_t>(kx + 1)];
-        const auto sobel_x_idx = static_cast<size_t>(kx + 1);
-        const auto sobel_y_idx = static_cast<size_t>(ky + 1);
+        const auto sobel_x_idx = static_cast<size_t>(static_cast<long long>(kx) + 1LL);
+        const auto sobel_y_idx = static_cast<size_t>(static_cast<long long>(ky) + 1LL);
         const int kernel_value = kSobelYArray.at(sobel_y_idx).at(sobel_x_idx);
         sum += pixel * kernel_value;
       }
